# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document Search API",
        "description": "A powerful document search API engine using TNT Search for fast full-text search capabilities. This API supports multi-tenant workspaces, document upload, text extraction, and advanced search features.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search documents",
                "description": "Search for documents using full-text search with filters",
                "operationId": "searchDocuments",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "annual report"
                        }
                    },
                    {
                        "name": "file_type",
                        "in": "query",
                        "description": "Filter by file type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pdf",
                                "doc",
                                "docx",
                                "txt"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter documents from date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter documents to date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "results": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/SearchResult"
                                                            }
                                                        },
                                                        "pagination": {
                                                            "$ref": "#/components/schemas/Pagination"
                                                        },
                                                        "query": {
                                                            "type": "string"
                                                        },
                                                        "total_found": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/search/advanced": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Advanced document search",
                "description": "Perform advanced search with multiple filters and sorting options",
                "operationId": "advancedSearchDocuments",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "financial report growth"
                                    },
                                    "filters": {
                                        "properties": {
                                            "file_type": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pdf",
                                                        "doc",
                                                        "docx",
                                                        "txt"
                                                    ]
                                                }
                                            },
                                            "date_from": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "date_to": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "sort": {
                                        "type": "string",
                                        "default": "relevance",
                                        "enum": [
                                            "relevance",
                                            "date_desc",
                                            "date_asc",
                                            "title_asc",
                                            "title_desc",
                                            "size_desc",
                                            "size_asc"
                                        ]
                                    },
                                    "page": {
                                        "type": "integer",
                                        "default": 1,
                                        "minimum": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "default": 10,
                                        "maximum": 100,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Advanced search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "results": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/SearchResult"
                                                            }
                                                        },
                                                        "pagination": {
                                                            "$ref": "#/components/schemas/Pagination"
                                                        },
                                                        "query": {
                                                            "type": "string"
                                                        },
                                                        "filters": {
                                                            "type": "object"
                                                        },
                                                        "sort": {
                                                            "type": "string"
                                                        },
                                                        "total_found": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/search/suggestions": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get search suggestions",
                "description": "Get search term suggestions based on indexed content",
                "operationId": "getSearchSuggestions",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Partial search query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "annu"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of suggestions",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5,
                            "maximum": 20,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search suggestions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "suggestions": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "query": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SearchResult": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Annual Report 2024"
                    },
                    "snippet": {
                        "type": "string",
                        "example": "The company achieved remarkable <mark>growth</mark> in revenue..."
                    },
                    "file_type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 2048576
                    },
                    "formatted_file_size": {
                        "type": "string",
                        "example": "2 MB"
                    },
                    "original_filename": {
                        "type": "string",
                        "example": "annual_report_2024.pdf"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": 0.95
                    }
                },
                "type": "object"
            },
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "WorkspaceKey": {
                "type": "apiKey",
                "description": "Workspace API key for authentication",
                "name": "X-Workspace-Key",
                "in": "header"
            }
        }
    },
...
}